moved prizeArr and symbols out from script.js into new file values.js, also changed the bet money and number of plays to 10k. One change per commit is correct way.
