+++ values.js
@@ -1,0 +1,16 @@
+let symbols = [
+    [25, 'Circle'],
+    [20, 'Drop'], 
+    [20, 'Eye'],
+    [15, 'Triangle'],
+    [15, 'Square'],
+    [5, 'Pentagram']];
+//0-5
+
+let prizeArr = [
+    13, 33, 33, 105, 105, 8533
+];
+
+
+module.exports.symbols = symbols;
+module.exports.prizeArr = prizeArr;
